refactor: move readSubscripts back into model.js and use `read` as the general entry point for all tests
